backend/requirements.txt
python-dotenv
django==4.2.4
django-cors-headers
django-neomodel
djangorestframework
neo4j==5.11.0
neo4j-driver==4.4.10
neobolt==1.7.17
neomodel==5.0.1
pandas
requests==2.31.0
gunicorn